backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
cohere
qdrant-client
pypdf
python-multipart
